Removed hard-coded dependencies
Removed dependency of \App\Http\Controllers\Controller

If an app wasn't using App as its root namespace, this library was broken. There was also no point of extending Controller as its functionality wasn't used in TwoFactorController.

Replaced auth()->user()->id for auth()->id()

Makes accessing the ID of the user more generic. Doesn't depend on how the column was named in the database.